requirements.txt
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
coverage==4.5.2
defusedxml==0.5.0
Django==2.1.3
django-cors-headers==2.4.0
django-cors-middleware==1.3.1
django-extensions==2.1.4
djangorestframework==3.9.0
docopt==0.6.2
flake8==3.6.0
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==2.1.0
pep8==1.7.1
pluggy==0.8.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
py==1.7.0
pycodestyle==2.4.0
pyflakes==2.0.0
PyJWT==1.7.0
pylint==2.2.2
python3-openid==3.1.0
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
urllib3==1.24.1
wrapt==1.10.11
